split up auth file, added nav bar and routes
