+++ adskfaysufgyasd/src/components/about.js
@@ -1,0 +1,27 @@
+import React, { useState } from 'react';
+
+export default function About(){
+
+    return(
+
+        <div>
+            # Run-For-The-Hills
+
+            **Introduction**
+
+            Welcome to Run for the Hills, a dating app make specifically for hillbillies, hill folk, and all other hill peeps*.  Life is complicated, dating shouldn’t be!
+
+            *not associated with the spooky cancerous hill movie concerning eyes
+
+            **Usage**
+
+            On this cutting edge dating app, users can find their potential happily-forever-afters through filters reflecting life’s most important attributes, such as beard length, truck lift size, tractor brand, and whether or they or not they “got the horses in the yard.”
+
+            To create an account, users will setup a username and password.  Don’t worry, our top notch hashing and salting methods ensure client security and safety from dark web hackers.  They can upload a photo from their computer and then fill in their stats.  Found oil on your property?  Bought another Ford?  Lost a tooth?  No worries, users can always go back and edit their profile!
+
+            Premium users have access to view people who have already liked them, cutting down time to the barn by a striking 81%!  Others are gonna have to be patient and search the entire haystack to find their needle.
+        </div>
+
+    )
+    
+}
